(feat): Refactors file organization to mimic react.
