+++ react/essentials-implementation/react-dom/index.ts
@@ -1,0 +1,8 @@
+import render, { reRender } from "./renderer";
+
+const ReactDOM = {
+	render,
+	reRender
+}
+
+export default ReactDOM;
