+++ react/essentials-implementation/react/index.ts
@@ -1,0 +1,9 @@
+import createElement from './createElement';
+
+// Default imports used through 'import React from "../react"'.
+const React = {
+	createElement
+}
+
+export default React;
+export { }
